requirements.txt
# ── Core App ─────────────────────────────────────────────────────────────────
streamlit

# ── LLM / Ollama ────────────────────────────────────────────────────────────
ollama
langchain
langchain-core
langchain-classic
langchain-community
langchain-huggingface
langchain-ollama
langchain-text-splitters
langgraph
langgraph-checkpoint-sqlite

# ── Vector Store & Embeddings ────────────────────────────────────────────────
faiss-cpu
sentence-transformers
huggingface_hub
transformers
tokenizers

# ── PyTorch (CPU-only for installer; swap for CUDA if GPU available) ────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# ── Retrieval Sources ───────────────────────────────────────────────────────
tavily-python
wikipedia
arxiv>=2.4.0
duckduckgo-search
ddgs

# ── Gmail ────────────────────────────────────────────────────────────────────
langchain-google-community[gmail]

# ── Document Loaders ────────────────────────────────────────────────────────
pypdf>=6.7.0

# ── Export ───────────────────────────────────────────────────────────────────
fpdf2

# ── Calculator ───────────────────────────────────────────────────────────────
simpleeval

# ── Wolfram Alpha ───────────────────────────────────────────────────────────
wolframalpha

# ── Vision (Camera + Screenshot) ─────────────────────────────────────────────────
opencv-python
mss

# ── Voice Input (STT) ───────────────────────────────────────────────────────
sounddevice
faster-whisper
openwakeword

# ── System Tray / Launcher ──────────────────────────────────────────────────
pystray
Pillow
